Extracted ConnectionSearch overlay and TrainSearch sub overlay from LeftMenu #22
